rename file types for CLObjects
